Relicense library as BSD.
Since is a complete rewrite of the code in Ruby, there is
no need to retain the original copyright of GPL.

This will also help with integration and usage in other
implementations like Rubinius.

Did this change with permission of Park Heesob, who donated
this code and give the copyright to RubyInstaller (One-Click)
project.