Add the ability to set the readline buffer.
This is needed for more advanced kinds of completions. Suppose I want to
expand:
   sh ma
into
   show macro

The completion function passes only the last token of the line, In
the trepanning debuggers, I use Readline.line_buffer to get at the entire
line entered and Readline.line_buffer= to set it after expansion on
all tokens is done.

- Unrelated -

Use phasis68 patches for issue luislaverna 26 and 27