Fix rl_filename_completion_function bug [issue: 35]
@filename and @dirname were not being set correctly. This meant that
trying to complete on pathnames that included directories did not work.

For example:

  /tmp/<tab><tab>

would not return any results. With this fix in place, it returns the
files and directories in /tmp, as expected.

I added tests which test the FILENAME_COMPLETION_PROC, rather than
rl_filename_completion_function, since I imagine it is what most Ruby
programs will actually be calling.

Signed-off-by: Mark Somerville <mark@scottishclimbs.com>