Readline.completion_append_character(): really return nil
When RbReadline.rl_completion_append_character == ?\0, the nil simply
fell through and we returned ?\0 instead. Adding a 'return' fixes it.

Note that the other places with a bare 'nil' in a if block are correct,
as there are no more statements following it.

This patch also ports the test_completion_append_character test from
MRI.

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F17, MRI 1.8.7/1.9.3]
[ mark@scottishclimbs.com: ensure we return the character rather than
                           the ASCII value under 1.8 ]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>