Port GNU Readline-5.2 patch-004
Fixes issue #53.

I am mostly happy with this port, but did struggle to port two
(identical) C conditionals:

  if (current_line ... &&  ((nfd - new) <= prompt_last_invisible))

In Ruby, I have used:

  if current_line ... && nfd <= @prompt_last_invisible

The reason for this is that in C, nfd and new are both char*, but in
rb-readline, nfd is an integer and new is a string. Unfortunately, it's
not *exactly* obvious to me what (nfd - new) actually returns in C.
However, in Ruby, nfd is an Integer which describes the index of the
first difference in the new string, so this code does seem to make sense
to me.

The message from the GNU commit[1] is:

-----
Bug-Reported-by: Peter Volkov <torre_cremata@mail.ru>
Bug-Reference-ID: <1173636022.7039.36.camel@localhost>
Bug-Reference-URL:
http://lists.gnu.org/archive/html/bug-bash/2007-03/msg00039.html

Bug-Description:

When restoring the original prompt after finishing an incremental search,
bash sometimes places the cursor incorrectly if the primary prompt contains
invisible characters.
-----

1 - http://ftp.gnu.org/gnu/readline/readline-5.2-patches/readline52-004

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F17, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>