Fixes #69: display issue scrolling past long lines
When scrolling through the history, if we encountered a line that is
longer than the terminal width, the next (shorter) line in the history
would cause a display problem. Often meant that the prompt would
disappear, but sometimes the entire line would be blank.

Any off-by-one error was causing this.

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F17, MRI 1.8.7]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>