Fixes #73: exception when scolling through long lines
If a line is in the history that occupies more than 2 lines on the
screen, scrolling through the history to it would raise an exception.

This patch removes some lines that are not present in GNU C Readline. I
can't find out or work out why these lines were there. Unless we get
some regression bug reports, it appears that they were causing more
problems than they solved.

Tested-by: Mark Somerville <ms@pccl.info> [F17, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>