Unload Readline before loading rb-readline
Requiring rb-readline after GNU Readline was giving "already initialized
constant" warnings. To prevent possible conflicts arising between the
GNU Readline C extension module and the Readline module defined by
rb-readline, this patch removes the Readline constant on require.

This ensures that we're using rb-readline in a clean state, but is not
without a different set of problems - any calls that had been made to
methods like Readline.completion_proc will need to be made again, as
will things like monkey-patches to the Readline module. I think that in
most real-world usage this will not be a problem as it would seem usual
for rb-readline to be called near the start of execution. Also, I can't
think of an alternative way of fixing this.

Closes #49.

Tested-by: Mark Somerville <ms@pccl.info> [F16/F17, MRI 1.9.3/2.0.0-preview2]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>