Add condtion missed when porting Readline-5.2 patch-013
In 15790be7e Readline-5.2 patch-013 was ported, but a line was missed.

Unfortunately, even with this line, some problems are present. For
example, when starting a new Urchin shell with invisible characters in
the prompt, the cursor is incorrectly placed at the leftmost character
on the screen.

It is likely that this patch and 15790be7e will be reverted, but it is
included here for completeness.

Tested-by: Mark Somerville <mark@jellyjelly.net> [Arch Linux, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@jellyjelly.net>