Revert "Port GNU Readline-5.2 patch-013"
This patch appeared to cause more problems than it solved. For example,
when starting a new Urchin shell with invisible characters in the
prompt, the cursor is incorrectly placed at the leftmost character on
the screen.

Since no known rb-readline users have reported problems like the one
that the original patch (15790be7e30) solves, reverting this makes sense
at the current time.

This reverts two commits:

  710324b20f272b0645f0093674c82223b2713635
  15790be7e30d80d38572041698d0114bf4315bff