Marshal HANDLE correctly on Win64
HANDLE is a void* and must use Fiddle::TYPE_VOIDP, not TYPE_LONG. Custom
marshalling of pointer parameters is removed as it only works for string
values. But such pointer parameters can also be Fixnums with the pointer
address (in case of HANDLEs). Fiddle already takes care of this and also
handles nil values correctly.